module for pump control
