+++ version.txt
@@ -1,1 +1,1 @@
-v1.1.24
+v1.1.25
